+++ node/src/components/consensus/config.rs
@@ -7,7 +7,12 @@ use casper_types::{PublicKey, SecretKey};
 
 use crate::{
     components::consensus::{
-        era_supervisor::PAST_OPEN_ERAS, protocols::highway::config::Config as HighwayConfig, EraId,
+        era_supervisor::PAST_OPEN_ERAS,
+        protocols::{
+            highway::config::Config as HighwayConfig,
+            simple_consensus::config::Config as SimpleConsensusConfig,
+        },
+        EraId,
     },
     types::Chainspec,
     utils::{External, LoadError, Loadable},
@@ -20,15 +25,22 @@ use crate::{
 pub(crate) struct Config {
     /// Path to secret key file.
     pub(crate) secret_key_path: External,
+    /// The maximum number of blocks by which execution is allowed to lag behind finalization.
+    /// If it is more than that, consensus will pause, and resume once the executor has caught up.
+    pub max_execution_delay: u64,
     /// Highway-specific node configuration.
     pub(crate) highway: HighwayConfig,
+    /// SimpleConsensus-specific node configuration.
+    pub(crate) simple_consensus: SimpleConsensusConfig,
 }
 
 impl Default for Config {
     fn default() -> Self {
         Config {
             secret_key_path: External::Missing,
+            max_execution_delay: 3,
             highway: HighwayConfig::default(),
+            simple_consensus: SimpleConsensusConfig::default(),
         }
     }
 }