+++ node/src/types/chainspec/core_config.rs
@@ -1,8 +1,15 @@
 use datasize::DataSize;
 use num::rational::Ratio;
 #[cfg(test)]
-use rand::Rng;
-use serde::{Deserialize, Serialize};
+use rand::{
+    distributions::{Distribution, Standard},
+    Rng,
+};
+use serde::{
+    de::{Deserializer, Error as DeError},
+    Deserialize, Serialize, Serializer,
+};
+use tracing::error;
 
 use casper_types::bytesrepr::{self, FromBytes, ToBytes};
 #[cfg(test)]
@@ -16,7 +23,10 @@ use crate::types::TimeDiff;
 pub struct CoreConfig {
     pub(crate) era_duration: TimeDiff,
     pub(crate) minimum_era_height: u64,
+    pub(crate) minimum_block_time: TimeDiff,
     pub(crate) validator_slots: u32,
+    #[data_size(skip)]
+    pub(crate) finality_threshold_fraction: Ratio<u64>,
     /// Number of eras before an auction actually defines the set of validators.
     /// If you bond with a sufficient bid in era N, you will be a validator in era N +
     /// auction_delay + 1
@@ -36,6 +46,24 @@ pub struct CoreConfig {
     pub(crate) minimum_delegation_amount: u64,
     /// Enables strict arguments checking when calling a contract.
     pub(crate) strict_argument_checking: bool,
+    /// Which consensus protocol to use.
+    pub(crate) consensus_protocol: ConsensusProtocolName,
+}
+
+impl CoreConfig {
+    /// Checks whether the values set in the config make sense and returns `false` if they don't.
+    pub(super) fn is_valid(&self) -> bool {
+        if self.finality_threshold_fraction <= Ratio::new(0, 1)
+            || self.finality_threshold_fraction >= Ratio::new(1, 1)
+        {
+            error!(
+                ftf = %self.finality_threshold_fraction,
+                "finality threshold fraction is not in the range (0, 1)",
+            );
+            return false;
+        }
+        true
+    }
 }
 
 #[cfg(test)]
@@ -44,7 +72,9 @@ impl CoreConfig {
     pub fn random(rng: &mut TestRng) -> Self {
         let era_duration = TimeDiff::from(rng.gen_range(600_000..604_800_000));
         let minimum_era_height = rng.gen_range(5..100);
-        let validator_slots = rng.gen();
+        let minimum_block_time = TimeDiff::from(rng.gen_range(1_000..60_000));
+        let validator_slots = rng.gen_range(1..10_000);
+        let finality_threshold_fraction = Ratio::new(rng.gen_range(1..100), 100);
         let auction_delay = rng.gen::<u32>() as u64;
         let locked_funds_period = TimeDiff::from(rng.gen_range(600_000..604_800_000));
         let unbonding_delay = rng.gen_range(1..1_000_000_000);
@@ -56,11 +86,14 @@ impl CoreConfig {
         let max_runtime_call_stack_height = rng.gen();
         let minimum_delegation_amount = rng.gen::<u32>() as u64;
         let strict_argument_checking = rng.gen();
+        let consensus_protocol = rng.gen();
 
         CoreConfig {
             era_duration,
             minimum_era_height,
+            minimum_block_time,
             validator_slots,
+            finality_threshold_fraction,
             auction_delay,
             locked_funds_period,
             unbonding_delay,
@@ -69,6 +102,7 @@ impl CoreConfig {
             max_runtime_call_stack_height,
             minimum_delegation_amount,
             strict_argument_checking,
+            consensus_protocol,
         }
     }
 }
@@ -78,7 +112,9 @@ impl ToBytes for CoreConfig {
         let mut buffer = bytesrepr::allocate_buffer(self)?;
         buffer.extend(self.era_duration.to_bytes()?);
         buffer.extend(self.minimum_era_height.to_bytes()?);
+        buffer.extend(self.minimum_block_time.to_bytes()?);
         buffer.extend(self.validator_slots.to_bytes()?);
+        buffer.extend(self.finality_threshold_fraction.to_bytes()?);
         buffer.extend(self.auction_delay.to_bytes()?);
         buffer.extend(self.locked_funds_period.to_bytes()?);
         buffer.extend(self.unbonding_delay.to_bytes()?);
@@ -87,13 +123,16 @@ impl ToBytes for CoreConfig {
         buffer.extend(self.max_runtime_call_stack_height.to_bytes()?);
         buffer.extend(self.minimum_delegation_amount.to_bytes()?);
         buffer.extend(self.strict_argument_checking.to_bytes()?);
+        buffer.extend(self.consensus_protocol.to_bytes()?);
         Ok(buffer)
     }
 
     fn serialized_length(&self) -> usize {
         self.era_duration.serialized_length()
             + self.minimum_era_height.serialized_length()
+            + self.minimum_block_time.serialized_length()
             + self.validator_slots.serialized_length()
+            + self.finality_threshold_fraction.serialized_length()
             + self.auction_delay.serialized_length()
             + self.locked_funds_period.serialized_length()
             + self.unbonding_delay.serialized_length()
@@ -102,6 +141,7 @@ impl ToBytes for CoreConfig {
             + self.max_runtime_call_stack_height.serialized_length()
             + self.minimum_delegation_amount.serialized_length()
             + self.strict_argument_checking.serialized_length()
+            + self.consensus_protocol.serialized_length()
     }
 }
 
@@ -109,7 +149,9 @@ impl FromBytes for CoreConfig {
     fn from_bytes(bytes: &[u8]) -> Result<(Self, &[u8]), bytesrepr::Error> {
         let (era_duration, remainder) = TimeDiff::from_bytes(bytes)?;
         let (minimum_era_height, remainder) = u64::from_bytes(remainder)?;
+        let (minimum_block_time, remainder) = TimeDiff::from_bytes(remainder)?;
         let (validator_slots, remainder) = u32::from_bytes(remainder)?;
+        let (finality_threshold_fraction, remainder) = Ratio::<u64>::from_bytes(remainder)?;
         let (auction_delay, remainder) = u64::from_bytes(remainder)?;
         let (locked_funds_period, remainder) = TimeDiff::from_bytes(remainder)?;
         let (unbonding_delay, remainder) = u64::from_bytes(remainder)?;
@@ -118,10 +160,13 @@ impl FromBytes for CoreConfig {
         let (max_runtime_call_stack_height, remainder) = u32::from_bytes(remainder)?;
         let (minimum_delegation_amount, remainder) = u64::from_bytes(remainder)?;
         let (strict_argument_checking, remainder) = bool::from_bytes(remainder)?;
+        let (consensus_protocol, remainder) = ConsensusProtocolName::from_bytes(remainder)?;
         let config = CoreConfig {
             era_duration,
             minimum_era_height,
+            minimum_block_time,
             validator_slots,
+            finality_threshold_fraction,
             auction_delay,
             locked_funds_period,
             unbonding_delay,
@@ -130,8 +175,78 @@ impl FromBytes for CoreConfig {
             max_runtime_call_stack_height,
             minimum_delegation_amount,
             strict_argument_checking,
+            consensus_protocol,
         };
         Ok((config, remainder))
+    }
+}
+
+#[derive(Copy, Clone, DataSize, PartialEq, Eq, Debug)]
+pub(crate) enum ConsensusProtocolName {
+    Highway,
+    Simple,
+}
+
+impl Serialize for ConsensusProtocolName {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        match self {
+            ConsensusProtocolName::Highway => "Highway",
+            ConsensusProtocolName::Simple => "Simple",
+        }
+        .serialize(serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for ConsensusProtocolName {
+    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
+        match String::deserialize(deserializer)?.to_lowercase().as_str() {
+            "highway" => Ok(ConsensusProtocolName::Highway),
+            "simple" => Ok(ConsensusProtocolName::Simple),
+            _ => Err(DeError::custom("unknown consensus protocol name")),
+        }
+    }
+}
+
+const CONSENSUS_HIGHWAY_TAG: u8 = 0;
+const CONSENSUS_SIMPLE_TAG: u8 = 1;
+
+impl ToBytes for ConsensusProtocolName {
+    fn to_bytes(&self) -> Result<Vec<u8>, bytesrepr::Error> {
+        let tag = match self {
+            ConsensusProtocolName::Highway => CONSENSUS_HIGHWAY_TAG,
+            ConsensusProtocolName::Simple => CONSENSUS_SIMPLE_TAG,
+        };
+        Ok(vec![tag])
+    }
+
+    fn serialized_length(&self) -> usize {
+        1
+    }
+}
+
+impl FromBytes for ConsensusProtocolName {
+    fn from_bytes(bytes: &[u8]) -> Result<(Self, &[u8]), bytesrepr::Error> {
+        let (tag, remainder) = u8::from_bytes(bytes)?;
+        let name = match tag {
+            CONSENSUS_HIGHWAY_TAG => ConsensusProtocolName::Highway,
+            CONSENSUS_SIMPLE_TAG => ConsensusProtocolName::Simple,
+            _ => return Err(bytesrepr::Error::Formatting),
+        };
+        Ok((name, remainder))
+    }
+}
+
+#[cfg(test)]
+impl Distribution<ConsensusProtocolName> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ConsensusProtocolName {
+        if rng.gen() {
+            ConsensusProtocolName::Highway
+        } else {
+            ConsensusProtocolName::Simple
+        }
     }
 }
 
@@ -153,5 +268,25 @@ mod tests {
         let encoded = toml::to_string_pretty(&config).unwrap();
         let decoded = toml::from_str(&encoded).unwrap();
         assert_eq!(config, decoded);
+    }
+
+    #[test]
+    fn should_validate_for_finality_threshold() {
+        let mut rng = crate::new_rng();
+        let mut config = CoreConfig::random(&mut rng);
+        // Should be valid for FTT > 0 and < 1.
+        config.finality_threshold_fraction = Ratio::new(1, u64::MAX);
+        assert!(config.is_valid());
+        config.finality_threshold_fraction = Ratio::new(u64::MAX - 1, u64::MAX);
+        assert!(config.is_valid());
+        // Should be invalid for FTT == 0 or >= 1.
+        config.finality_threshold_fraction = Ratio::new(0, 1);
+        assert!(!config.is_valid());
+        config.finality_threshold_fraction = Ratio::new(1, 1);
+        assert!(!config.is_valid());
+        config.finality_threshold_fraction = Ratio::new(u64::MAX, u64::MAX);
+        assert!(!config.is_valid());
+        config.finality_threshold_fraction = Ratio::new(u64::MAX, u64::MAX - 1);
+        assert!(!config.is_valid());
     }
 }