+++ node/src/types/chainspec/highway_config.rs
@@ -15,8 +15,6 @@ use crate::types::TimeDiff;
 // Disallow unknown fields to ensure config files and command-line overrides contain valid keys.
 #[serde(deny_unknown_fields)]
 pub(crate) struct HighwayConfig {
-    #[data_size(skip)]
-    pub(crate) finality_threshold_fraction: Ratio<u64>,
     pub(crate) minimum_round_exponent: u8,
     pub(crate) maximum_round_exponent: u8,
     /// The factor by which rewards for a round are multiplied if the greatest summit has ≤50%
@@ -33,16 +31,6 @@ impl HighwayConfig {
                 min = %self.minimum_round_exponent,
                 max = %self.maximum_round_exponent,
                 "minimum round exponent is greater than the maximum round exponent",
-            );
-            return false;
-        }
-
-        if self.finality_threshold_fraction <= Ratio::new(0, 1)
-            || self.finality_threshold_fraction >= Ratio::new(1, 1)
-        {
-            error!(
-                ftf = %self.finality_threshold_fraction,
-                "finality threshold fraction is not in the range (0, 1)",
             );
             return false;
         }
@@ -68,13 +56,11 @@ impl HighwayConfig {
 impl HighwayConfig {
     /// Generates a random instance using a `TestRng`.
     pub fn random(rng: &mut TestRng) -> Self {
-        let finality_threshold_fraction = Ratio::new(rng.gen_range(1..100), 100);
         let minimum_round_exponent = rng.gen_range(0..16);
         let maximum_round_exponent = rng.gen_range(16..22);
         let reduced_reward_multiplier = Ratio::new(rng.gen_range(0..10), 10);
 
         HighwayConfig {
-            finality_threshold_fraction,
             minimum_round_exponent,
             maximum_round_exponent,
             reduced_reward_multiplier,
@@ -85,7 +71,6 @@ impl HighwayConfig {
 impl ToBytes for HighwayConfig {
     fn to_bytes(&self) -> Result<Vec<u8>, bytesrepr::Error> {
         let mut buffer = bytesrepr::allocate_buffer(self)?;
-        buffer.extend(self.finality_threshold_fraction.to_bytes()?);
         buffer.extend(self.minimum_round_exponent.to_bytes()?);
         buffer.extend(self.maximum_round_exponent.to_bytes()?);
         buffer.extend(self.reduced_reward_multiplier.to_bytes()?);
@@ -93,8 +78,7 @@ impl ToBytes for HighwayConfig {
     }
 
     fn serialized_length(&self) -> usize {
-        self.finality_threshold_fraction.serialized_length()
-            + self.minimum_round_exponent.serialized_length()
+        self.minimum_round_exponent.serialized_length()
             + self.maximum_round_exponent.serialized_length()
             + self.reduced_reward_multiplier.serialized_length()
     }
@@ -102,12 +86,10 @@ impl ToBytes for HighwayConfig {
 
 impl FromBytes for HighwayConfig {
     fn from_bytes(bytes: &[u8]) -> Result<(Self, &[u8]), bytesrepr::Error> {
-        let (finality_threshold_fraction, remainder) = Ratio::<u64>::from_bytes(bytes)?;
-        let (minimum_round_exponent, remainder) = u8::from_bytes(remainder)?;
+        let (minimum_round_exponent, remainder) = u8::from_bytes(bytes)?;
         let (maximum_round_exponent, remainder) = u8::from_bytes(remainder)?;
         let (reduced_reward_multiplier, remainder) = Ratio::<u64>::from_bytes(remainder)?;
         let config = HighwayConfig {
-            finality_threshold_fraction,
             minimum_round_exponent,
             maximum_round_exponent,
             reduced_reward_multiplier,
@@ -149,28 +131,6 @@ mod tests {
 
         // Should be invalid for round exponents where min > max.
         highway_config.minimum_round_exponent = highway_config.maximum_round_exponent + 1;
-        assert!(!highway_config.is_valid());
-    }
-
-    #[test]
-    fn should_validate_for_finality_threshold() {
-        let mut rng = crate::new_rng();
-        let mut highway_config = HighwayConfig::random(&mut rng);
-
-        // Should be valid for FTT > 0 and < 1.
-        highway_config.finality_threshold_fraction = Ratio::new(1, u64::MAX);
-        assert!(highway_config.is_valid());
-        highway_config.finality_threshold_fraction = Ratio::new(u64::MAX - 1, u64::MAX);
-        assert!(highway_config.is_valid());
-
-        // Should be invalid for FTT == 0 or >= 1.
-        highway_config.finality_threshold_fraction = Ratio::new(0, 1);
-        assert!(!highway_config.is_valid());
-        highway_config.finality_threshold_fraction = Ratio::new(1, 1);
-        assert!(!highway_config.is_valid());
-        highway_config.finality_threshold_fraction = Ratio::new(u64::MAX, u64::MAX);
-        assert!(!highway_config.is_valid());
-        highway_config.finality_threshold_fraction = Ratio::new(u64::MAX, u64::MAX - 1);
         assert!(!highway_config.is_valid());
     }