+++ node/CHANGELOG.md
@@ -28,6 +28,11 @@ All notable changes to this project will be documented in this file.  The format
 * Add run-mode field to the `/status` endpoint and the `info_get_status` JSON-RPC.
 * Add new REST `/chainspec` and JSON-RPC `info_get_chainspec` endpoints that return the raw bytes of the `chainspec.toml`, `accounts.toml` and `global_state.toml` files as read at node startup.
 * Add a new parameter to `info_get_deploys` JSON-RPC, `finalized_approvals` - controlling whether the approvals returned with the deploy should be the ones originally received by the node, or overridden by the approvals that were finalized along with the deploy.
+* Make consensus settings non-optional. A value 0 disables them.
+* Add a `[consensus.simple_consensus]` section to `config.toml` for the simple consensus protocol.
+* Add a `consensus_protocol` setting to the chainspec to choose a consensus protocol, and a `minimum_block_time` setting for the minimum difference between a block's timestamp and its child's.
+* Move `finality_threshold_fraction` from the `highway` to the `core` section in the chainspec.
+* Move `max_execution_delay` from the `highway` to the `consensus` section in the `config.toml`.
 
 ### Changed
 * Detection of a crash no longer triggers DB integrity checks to run on node start; the checks can be triggered manually instead.